HDFS-5906. Fixing findbugs and javadoc warnings in the HDFS-5698 branch. Contributed by Haohui Mai.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/common/branches/HDFS-5698@1565765 13f79535-47bb-0310-9956-ffa450edef68

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/tools/offlineImageViewer/FileDistributionCalculator.java
@@ -129,7 +129,7 @@ final class FileDistributionCalculator {
         totalSpace += fileSize;
 
         int bucket = fileSize > maxSize ? distribution.length - 1 : (int) Math
-            .ceil(fileSize / steps);
+            .ceil((double)fileSize / steps);
         ++distribution[bucket];
 
       } else if (p.getType() == INodeSection.INode.Type.DIRECTORY) {
